draft: introduce pageable responses in API calls
The API calls in the application have been updated to return pageable responses. The size and page parameters were added to the request instead of limit and offset. The response data now contains a pageable object, with the actual content extracted and used in the application. Also, a new function for loading more entries has been added.