Highlight current route in sidebar menu
The sidebar menu now adapts to user navigation by highlighting the current route. This was achieved by importing the Vue router in "SidebarMenuItem.vue" and "SidebarFeedItem.vue" components and creating a helper function. The helper function, 'isCurrentRoute', verifies if the current route path matches the prop 'to'. It then updates the class attribute dynamically to add an additional border effect.