draft: add LogoutView and update routing
A LogoutView is created in the authentication folder, with functionality to log out the user and redirect them to the login page. The routing file has been updated to include a path for logout, ensuring users will be directed to the LogoutView when accessing the '/logout' path. The location of the LoginView has also been moved to the authentication folder.