draft: add 'Request Account' feature
A new feature allowing users to request for new account creation has been implemented. This includes setting up a new route '/request-account', creating a new RequestAccountView.vue file which contains the form for account requests and updating the authentication store to handle account request operations. A minor change was also made to the LoginView.vue to correct user redirections when already logged in.